+++ src/debug/internal.ts
@@ -1,0 +1,2 @@
+export * from "./AffiseDebugApi";
+export * from "./AffiseDebug";