+++ src/module/appsflyer/AffiseAppsFlyer.ts
@@ -1,0 +1,19 @@
+import { AffiseNative } from "../../native/AffiseNative";
+import { AffiseModuleAppsFlyerApi } from "./AffiseModuleAppsFlyerApi";
+
+export class AffiseAppsFlyer implements AffiseModuleAppsFlyerApi { 
+    private native: AffiseNative
+
+    constructor(native: AffiseNative) { 
+        this.native = native
+    }
+    
+    /**
+     * Send Appflyer event data to affise
+     * @param eventName event name
+     * @param eventValues event data
+     */
+    logEvent(eventName: string, eventValues: Record<string, any>) {
+        this.native.logEvent(eventName, eventValues);
+    }
+}