+++ src/module/appsflyer/AffiseModuleAppsFlyerApi.ts
@@ -1,0 +1,8 @@
+export interface AffiseModuleAppsFlyerApi { 
+    /**
+     * Send Appflyer event data to affise
+     * @param eventName event name
+     * @param eventValues event data
+     */
+    logEvent(eventName: string, eventValues: Record<string, any>): void;
+}