+++ src/module/attribution/AffiseAndroidApi.ts
@@ -1,0 +1,8 @@
+export interface AffiseAndroidApi { 
+    /**
+     * Erases all user data from mobile and sends [GDPREvent]
+     */
+    forget(userData: string): void;
+
+    crashApplication(): void;
+}